Enhance facility information display when closed
- Updated map popups to show opening hours for closed facilities.
- Modified left sidebar accordion to display opening hours for closed academic buildings.
- Created a shared utility file for formatTime and formatDuration functions to eliminate code duplication.
