+++ utils/format.ts
@@ -1,0 +1,16 @@
+export const formatTime = (time: string | undefined): string => {
+  if (!time) return "";
+  const [hours, minutes] = time.split(":");
+  const hour = parseInt(hours, 10);
+  const ampm = hour >= 12 ? "PM" : "AM";
+  const hour12 = hour % 12 || 12;
+  return `${hour12}:${minutes} ${ampm}`;
+};
+
+export const formatDuration = (minutes: number | undefined): string => {
+  if (!minutes) return "";
+  if (minutes < 60) return `${Math.floor(minutes)} min`;
+  const hours = Math.floor(minutes / 60);
+  const remainingMinutes = Math.floor(minutes % 60);
+  return remainingMinutes > 0 ? `${hours}h ${remainingMinutes}m` : `${hours}h`;
+};
